Reorganise fixtures and thest the login
